fix: expected packages test
