+++ packages/package.json
@@ -49,6 +49,7 @@
 		"babel-jest": "^29.7.0",
 		"copyfiles": "^2.4.1",
 		"jest": "^29.7.0",
+		"jest-mock-extended": "^3.0.5",
 		"json-schema-to-typescript": "^10.1.5",
 		"lerna": "^6.6.2",
 		"nodemon": "^2.0.22",
